requirements.txt
# Dipendenze principali
pandas>=1.3.0
pyyaml>=6.0
streamlit>=1.10.0
plotly>=5.5.0
openpyxl>=3.0.9  # Per il supporto Excel in pandas

# Dipendenze opzionali
matplotlib>=3.5.0  # Per grafici aggiuntivi
numpy>=1.20.0     # Richiesto da pandas e plotly